requirements.txt
asgiref == 3.8.1
Django == 5.1.1
djangorestframework == 3.16.0
dnspython == 2.7.0
pip==24.2
psycopg2-binary == 2.9.10
pymongo == 4.12.0
python-dotenv == 1.1.0
setuptools == 75.1.0
sqlparse == 0.5.1
wheel == 0.44.0